chore: import flyway to migration db - add several enums to fit db - change related entities properties type - remove several useless code

+++ waterfun-admin-service/src/main/java/org/waterwood/waterfunadminservice/api/request/user/CreateNewUserReq.java
@@ -21,5 +21,5 @@ public class CreateNewUserReq {
     @StrongPassword
     private String password;
     @NotNull
-    private Short userType;
+    private Byte userType;
 }
