+++ waterfun-service-core/src/main/java/org/waterwood/waterfunservicecore/entity/audit/AuditLogStatusTypeConverter.java
@@ -4,14 +4,14 @@ import jakarta.persistence.AttributeConverter;
 import jakarta.persistence.Converter;
 
 @Converter(autoApply = true)
-public class AuditLogStatusTypeConverter implements AttributeConverter<AuditLogStatusType, Short> {
+public class AuditLogStatusTypeConverter implements AttributeConverter<AuditLogStatusType, Byte> {
     @Override
-    public Short convertToDatabaseColumn(AuditLogStatusType attribute) {
+    public Byte convertToDatabaseColumn(AuditLogStatusType attribute) {
         return attribute == null ? null : attribute.getValue();
     }
 
     @Override
-    public AuditLogStatusType convertToEntityAttribute(Short dbData) {
+    public AuditLogStatusType convertToEntityAttribute(Byte dbData) {
         return dbData == null ? null : AuditLogStatusType.fromValue(dbData);
     }
 }
