+++ waterfun-service-core/src/main/java/org/waterwood/waterfunservicecore/entity/audit/RejectTypeConverter.java
@@ -4,14 +4,14 @@ import jakarta.persistence.AttributeConverter;
 import jakarta.persistence.Converter;
 
 @Converter(autoApply = true)
-public class RejectTypeConverter implements AttributeConverter<AuditType, Short> {
+public class RejectTypeConverter implements AttributeConverter<AuditType, Byte> {
     @Override
-    public Short convertToDatabaseColumn(AuditType attribute) {
-        return attribute == null ? null : (short) attribute.getCode();
+    public Byte convertToDatabaseColumn(AuditType attribute) {
+        return attribute == null ? null : attribute.getCode();
     }
 
     @Override
-    public AuditType convertToEntityAttribute(Short dbData) {
-        return dbData == null ? null : AuditType.fromCode(dbData.intValue());
+    public AuditType convertToEntityAttribute(Byte dbData) {
+        return dbData == null ? null : AuditType.fromCode(dbData);
     }
 }
