+++ waterfun-service-core/src/main/java/org/waterwood/waterfunservicecore/entity/content/BannerPositionConvertor.java
@@ -4,15 +4,15 @@ import jakarta.persistence.AttributeConverter;
 import jakarta.persistence.Converter;
 
 @Converter(autoApply = true)
-public class BannerPositionConvertor implements AttributeConverter<BannerPosition, Short> {
+public class BannerPositionConvertor implements AttributeConverter<BannerPosition, Byte> {
 
     @Override
-    public Short convertToDatabaseColumn(BannerPosition attribute) {
+    public Byte convertToDatabaseColumn(BannerPosition attribute) {
         return attribute == null ? 1 : attribute.getCode();
     }
 
     @Override
-    public BannerPosition convertToEntityAttribute(Short dbData) {
+    public BannerPosition convertToEntityAttribute(Byte dbData) {
         return BannerPosition.fromCode(dbData);
     }
 }
