+++ waterfun-service-core/src/main/java/org/waterwood/waterfunservicecore/entity/perm/PermissionTypeConverter.java
@@ -4,15 +4,15 @@ import jakarta.persistence.AttributeConverter;
 import jakarta.persistence.Converter;
 
 @Converter(autoApply = true)
-public class PermissionTypeConverter implements AttributeConverter<PermissionType, Short> {
+public class PermissionTypeConverter implements AttributeConverter<PermissionType, Byte> {
 
     @Override
-    public Short convertToDatabaseColumn(PermissionType attribute) {
+    public Byte convertToDatabaseColumn(PermissionType attribute) {
         return attribute == null ? null : attribute.getValue();
     }
 
     @Override
-    public PermissionType convertToEntityAttribute(Short dbData) {
+    public PermissionType convertToEntityAttribute(Byte dbData) {
         return dbData == null ? null : PermissionType.fromValue(dbData);
     }
 }
