+++ waterfun-service-core/src/main/java/org/waterwood/waterfunservicecore/entity/post/PostType.java
@@ -7,12 +7,12 @@ public enum PostType {
     COMMON(0),
     NOTICE(1),;
 
-    private final short value;
+    private final byte value;
     private PostType(int value) {
-        this.value = (short) value;
+        this.value = (byte) value;
     }
 
-    public static PostType fromValue(short value) {
+    public static PostType fromValue(Byte value) {
         for (PostType postType : PostType.values()) {
             if (postType.value == value) {
                 return postType;
